@@ -6,30 +6,35 @@
     "jtl",
     "Connector"
   ],
-  "homepage": "http://www.jtl-software.de",
+  "homepage": "https://www.jtl-software.de",
   "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "JTL Software GmbH",
       "email": "info@jtl-software.de",
-      "homepage": "http://www.jtl-software.de",
+      "homepage": "https://www.jtl-software.de",
       "role": "Founder"
     }
   ],
   "minimum-stability": "dev",
   "prefer-stable": true,
   "config": {
-    "platform": {
-      "php": "7.1.3"
-    },
     "vendor-dir": "lib"
   },
   "require": {
-    "php": ">=7.1.3",
-    "jtl/connector": "^3.2.0",
+    "php": ">= 8.0",
+    "jtl/connector": "^5.2.10",
     "symfony/yaml": "^3.4",
     "ext-json": "*",
-    "phing/phing": "^2.16"
+    "phing/phing": "^2.16",
+    "psr/cache": "^2",
+    "monolog/monolog": "1.*",
+    "psr/log": "^1",
+    "symfony/http-foundation": "^4.4",
+    "jms/serializer": ">=3.3 <=3.24",
+    "doctrine/lexer": "^1",
+    "symfony/service-contracts": "^2.5",
+    "symfony/event-dispatcher-contracts": "^1.2"
   },
   "autoload": {
     "psr-4": {
@@ -41,7 +46,6 @@
     }
   },
   "require-dev": {
-    "jtl/connector-integration-tests": "dev-develop",
     "friendsofphp/php-cs-fixer": "^2.16",
     "symfony/polyfill-php80": "<=1.18.1"
   }
